Refactor person state
