+++ plugin-server/src/worker/ingestion/event-pipeline/processPersonsStep.ts
@@ -3,7 +3,10 @@ import { DateTime } from 'luxon'
 
 import { Person, Team } from '~/types'
 
-import { PersonState } from '../persons/person-state'
+import { PersonContext } from '../persons/person-context'
+import { PersonEventProcessor } from '../persons/person-event-processor'
+import { PersonMergeService } from '../persons/person-merge.service'
+import { PersonPropertyService } from '../persons/person-property.service'
 import { PersonsStoreForBatch } from '../persons/persons-store-for-batch'
 import { EventPipelineRunner } from './runner'
 
@@ -15,7 +18,7 @@ export async function processPersonsStep(
     processPerson: boolean,
     personStoreBatch: PersonsStoreForBatch
 ): Promise<[PluginEvent, Person, Promise<void>]> {
-    const [person, kafkaAck] = await new PersonState(
+    const context = new PersonContext(
         event,
         team,
         String(event.distinct_id),
@@ -25,7 +28,14 @@ export async function processPersonsStep(
         personStoreBatch,
         runner.hub.PERSON_JSONB_SIZE_ESTIMATE_ENABLE,
         runner.hub.PERSON_PROPERTY_JSONB_UPDATE_OPTIMIZATION
-    ).update()
+    )
+
+    const processor = new PersonEventProcessor(
+        context,
+        new PersonPropertyService(context),
+        new PersonMergeService(context)
+    )
+    const [person, kafkaAck] = await processor.processEvent()
 
     return [event, person, kafkaAck]
 }
