+++ plugin-server/tests/worker/ingestion/person-state.test.ts
@@ -1,8 +1,12 @@
 import { PluginEvent, Properties } from '@posthog/plugin-scaffold'
 import { DateTime } from 'luxon'
 
-import { TopicMessage } from '~/kafka/producer'
-import { MeasuringPersonsStoreForBatch } from '~/worker/ingestion/persons/measuring-person-store'
+import { TopicMessage } from '../../../src/kafka/producer'
+import { MeasuringPersonsStoreForBatch } from '../../../src/worker/ingestion/persons/measuring-person-store'
+import { PersonContext } from '../../../src/worker/ingestion/persons/person-context'
+import { PersonEventProcessor } from '../../../src/worker/ingestion/persons/person-event-processor'
+import { PersonMergeService } from '../../../src/worker/ingestion/persons/person-merge.service'
+import { PersonPropertyService } from '../../../src/worker/ingestion/persons/person-property.service'
 
 import {
     Database,
@@ -18,7 +22,6 @@ import { PostgresUse, TransactionClient } from '../../../src/utils/db/postgres'
 import { defaultRetryConfig } from '../../../src/utils/retries'
 import { UUIDT } from '../../../src/utils/utils'
 import { uuidFromDistinctId } from '../../../src/worker/ingestion/person-uuid'
-import { PersonState } from '../../../src/worker/ingestion/persons/person-state'
 import { delayUntilEventIngested } from '../../helpers/clickhouse'
 import { createOrganization, createTeam, fetchPostgresPersons, getTeam, insertRow } from '../../helpers/sql'
 
@@ -57,7 +60,7 @@ async function createPerson(
     return person
 }
 
-describe('PersonState.update()', () => {
+describe('PersonState.processEvent()', () => {
     let hub: Hub
 
     let teamId: number
@@ -106,7 +109,72 @@ describe('PersonState.update()', () => {
         await hub.db.clickhouseQuery('SYSTEM START MERGES')
     })
 
-    function personState(
+    function personProcessor(
+        event: Partial<PluginEvent>,
+        propertyService?: PersonPropertyService,
+        mergeService?: PersonMergeService,
+        customHub?: Hub,
+        processPerson = true,
+        timestampParam = timestamp,
+        team = mainTeam
+    ) {
+        const fullEvent = {
+            team_id: teamId,
+            properties: {},
+            ...event,
+        }
+
+        const personsStore = new MeasuringPersonsStoreForBatch(customHub ? customHub.db : hub.db)
+
+        const context = new PersonContext(
+            fullEvent as any,
+            team,
+            event.distinct_id!,
+            timestampParam,
+            processPerson,
+            customHub ? customHub.db.kafkaProducer : hub.db.kafkaProducer,
+            personsStore,
+            0
+        )
+        const processor = new PersonEventProcessor(
+            context,
+            propertyService ?? new PersonPropertyService(context),
+            mergeService ?? new PersonMergeService(context)
+        )
+        return processor
+    }
+
+    function personPropertyService(
+        event: Partial<PluginEvent>,
+        customHub?: Hub,
+        processPerson = true,
+        timestampParam = timestamp,
+        team = mainTeam,
+        updateIsIdentified: boolean = false
+    ) {
+        const fullEvent = {
+            team_id: teamId,
+            properties: {},
+            ...event,
+        }
+
+        const personsStore = new MeasuringPersonsStoreForBatch(customHub ? customHub.db : hub.db)
+
+        const context = new PersonContext(
+            fullEvent as any,
+            team,
+            event.distinct_id!,
+            timestampParam,
+            processPerson,
+            customHub ? customHub.db.kafkaProducer : hub.db.kafkaProducer,
+            personsStore,
+            0
+        )
+        context.updateIsIdentified = updateIsIdentified
+        return new PersonPropertyService(context)
+    }
+
+    function personMergeService(
         event: Partial<PluginEvent>,
         customHub?: Hub,
         processPerson = true,
@@ -120,7 +188,8 @@ describe('PersonState.update()', () => {
         }
 
         const personsStore = new MeasuringPersonsStoreForBatch(customHub ? customHub.db : hub.db)
-        return new PersonState(
+
+        const context = new PersonContext(
             fullEvent as any,
             team,
             event.distinct_id!,
@@ -130,6 +199,7 @@ describe('PersonState.update()', () => {
             personsStore,
             0
         )
+        return new PersonMergeService(context)
     }
 
     const sortPersons = (persons: InternalPerson[]) => persons.sort((a, b) => Number(a.id) - Number(b.id))
@@ -166,7 +236,7 @@ describe('PersonState.update()', () => {
         it('creates deterministic person uuids that are different between teams', async () => {
             const event_uuid = new UUIDT().toString()
             const primaryTeamId = teamId
-            const [personPrimaryTeam, kafkaAcks] = await personState({
+            const [personPrimaryTeam, kafkaAcks] = await personPropertyService({
                 event: '$pageview',
                 distinct_id: newUserDistinctId,
                 uuid: event_uuid,
@@ -175,7 +245,7 @@ describe('PersonState.update()', () => {
             const otherTeamId = await createTeam(hub.db.postgres, organizationId)
             const otherTeam = (await getTeam(hub, otherTeamId))!
             teamId = otherTeamId
-            const [personOtherTeam, kafkaAcksOther] = await personState(
+            const [personOtherTeam, kafkaAcksOther] = await personPropertyService(
                 {
                     event: '$pageview',
                     distinct_id: newUserDistinctId,
@@ -201,16 +271,18 @@ describe('PersonState.update()', () => {
 
             const hubParam = undefined
             const processPerson = false
-            const [fakePerson, kafkaAcks] = await personState(
+            const [fakePerson, kafkaAcks] = await personProcessor(
                 {
                     event: '$pageview',
                     distinct_id: newUserDistinctId,
                     uuid: event_uuid,
                     properties: { $set: { should_be_dropped: 100 } },
                 },
+                undefined,
+                undefined,
                 hubParam,
                 processPerson
-            ).update()
+            ).processEvent()
             await hub.db.kafkaProducer.flush()
             await kafkaAcks
 
@@ -248,7 +320,7 @@ describe('PersonState.update()', () => {
 
             const hubParam = undefined
             const processPerson = true
-            const [_person, kafkaAcks] = await personState(
+            const [_person, kafkaAcks] = await personProcessor(
                 {
                     event: '$identify',
                     distinct_id: oldUserDistinctId,
@@ -256,11 +328,13 @@ describe('PersonState.update()', () => {
                         $anon_distinct_id: 'old2',
                     },
                 },
+                undefined,
+                undefined,
                 hubParam,
                 processPerson
-            ).update()
+            ).processEvent()
 
-            const [_person2, kafkaAcks2] = await personState(
+            const [_person2, kafkaAcks2] = await personProcessor(
                 {
                     event: '$identify',
                     distinct_id: newUserDistinctId,
@@ -268,9 +342,11 @@ describe('PersonState.update()', () => {
                         $anon_distinct_id: 'new2',
                     },
                 },
+                undefined,
+                undefined,
                 hubParam,
                 processPerson
-            ).update()
+            ).processEvent()
 
             await hub.db.kafkaProducer.flush()
             await kafkaAcks
@@ -311,7 +387,7 @@ describe('PersonState.update()', () => {
 
             const hubParam = undefined
             let processPerson = true
-            const [_person, kafkaAcks] = await personState(
+            const [_person, kafkaAcks] = await personProcessor(
                 {
                     event: '$identify',
                     distinct_id: newUserDistinctId,
@@ -319,9 +395,11 @@ describe('PersonState.update()', () => {
                         $anon_distinct_id: oldUserDistinctId,
                     },
                 },
+                undefined,
+                undefined,
                 hubParam,
                 processPerson
-            ).update()
+            ).processEvent()
             await hub.db.kafkaProducer.flush()
             await kafkaAcks
 
@@ -330,17 +408,19 @@ describe('PersonState.update()', () => {
             processPerson = false
             const event_uuid = new UUIDT().toString()
             const timestampParam = timestamp.plus({ minutes: 5 }) // Event needs to happen after Person creation
-            const [fakePerson, kafkaAcks2] = await personState(
+            const [fakePerson, kafkaAcks2] = await personProcessor(
                 {
                     event: '$pageview',
                     distinct_id: newUserDistinctId,
                     uuid: event_uuid,
                     properties: { $set: { should_be_dropped: 100 } },
                 },
+                undefined,
+                undefined,
                 hubParam,
                 processPerson,
                 timestampParam
-            ).update()
+            ).processEvent()
             await hub.db.kafkaProducer.flush()
             await kafkaAcks2
 
@@ -372,7 +452,7 @@ describe('PersonState.update()', () => {
 
             const hubParam = undefined
             let processPerson = true
-            const [_person, kafkaAcks] = await personState(
+            const [_person, kafkaAcks] = await personProcessor(
                 {
                     event: '$identify',
                     distinct_id: newUserDistinctId,
@@ -380,9 +460,11 @@ describe('PersonState.update()', () => {
                         $anon_distinct_id: oldUserDistinctId,
                     },
                 },
+                undefined,
+                undefined,
                 hubParam,
                 processPerson
-            ).update()
+            ).processEvent()
             await hub.db.kafkaProducer.flush()
             await kafkaAcks
 
@@ -391,17 +473,19 @@ describe('PersonState.update()', () => {
             processPerson = false
             const event_uuid = new UUIDT().toString()
             const timestampParam = timestamp.plus({ minutes: 5 }) // Event needs to happen after Person creation
-            const [fakePerson, kafkaAcks2] = await personState(
+            const [fakePerson, kafkaAcks2] = await personProcessor(
                 {
                     event: '$pageview',
                     distinct_id: newUserDistinctId,
                     uuid: event_uuid,
                     properties: { $set: { should_be_dropped: 100 } },
                 },
+                undefined,
+                undefined,
                 hubParam,
                 processPerson,
                 timestampParam
-            ).update()
+            ).processEvent()
             await hub.db.kafkaProducer.flush()
             await kafkaAcks2
 
@@ -410,7 +494,7 @@ describe('PersonState.update()', () => {
 
         it('creates person if they are new', async () => {
             const event_uuid = new UUIDT().toString()
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personPropertyService({
                 event: '$pageview',
                 distinct_id: newUserDistinctId,
                 uuid: event_uuid,
@@ -446,12 +530,12 @@ describe('PersonState.update()', () => {
 
         it('does not attach existing person properties to $process_person_profile=false events', async () => {
             const originalEventUuid = new UUIDT().toString()
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personProcessor({
                 event: '$pageview',
                 distinct_id: newUserDistinctId,
                 uuid: originalEventUuid,
                 properties: { $set: { c: 420 } },
-            }).update()
+            }).processEvent()
             await hub.db.kafkaProducer.flush()
             await kafkaAcks
 
@@ -477,25 +561,27 @@ describe('PersonState.update()', () => {
 
             // OK, a person now exists with { c: 420 }, let's prove the properties come back out
             // of the DB.
-            const [personVerifyProps] = await personState({
+            const [personVerifyProps] = await personProcessor({
                 event: '$pageview',
                 distinct_id: newUserDistinctId,
                 uuid: new UUIDT().toString(),
                 properties: {},
-            }).update()
+            }).processEvent()
             expect(personVerifyProps.properties).toEqual({ $creator_event_uuid: originalEventUuid, c: 420 })
 
             // But they don't when $process_person_profile=false
-            const [processPersonFalseResult] = await personState(
+            const [processPersonFalseResult] = await personProcessor(
                 {
                     event: '$pageview',
                     distinct_id: newUserDistinctId,
                     uuid: new UUIDT().toString(),
                     properties: {},
                 },
+                undefined,
+                undefined,
                 hub,
                 false
-            ).update()
+            ).processEvent()
             expect(processPersonFalseResult.properties).toEqual({})
         })
 
@@ -517,7 +603,7 @@ describe('PersonState.update()', () => {
                 return Promise.resolve(undefined)
             })
 
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personPropertyService({
                 event: '$pageview',
                 distinct_id: newUserDistinctId,
             }).handleUpdate()
@@ -564,7 +650,7 @@ describe('PersonState.update()', () => {
                 return Promise.resolve(undefined)
             })
 
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personPropertyService({
                 event: '$pageview',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -598,7 +684,7 @@ describe('PersonState.update()', () => {
         })
 
         it('creates person with properties', async () => {
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personPropertyService({
                 event: '$pageview',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -639,7 +725,7 @@ describe('PersonState.update()', () => {
             await createPerson(hub, timestamp, { b: 3, c: 4, toString: {} }, {}, {}, teamId, null, false, newUserUuid, [
                 { distinctId: newUserDistinctId },
             ])
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personPropertyService({
                 event: '$pageview',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -677,7 +763,7 @@ describe('PersonState.update()', () => {
                 { distinctId: newUserDistinctId },
             ])
 
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personPropertyService({
                 event: event,
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -711,7 +797,7 @@ describe('PersonState.update()', () => {
             await createPerson(hub, timestamp, { $current_url: 123 }, {}, {}, teamId, null, false, newUserUuid, [
                 { distinctId: newUserDistinctId },
             ])
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personPropertyService({
                 event: '$pageview',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -753,7 +839,7 @@ describe('PersonState.update()', () => {
                 { distinctId: newUserDistinctId },
             ])
 
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personPropertyService({
                 event: '$set',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -787,7 +873,7 @@ describe('PersonState.update()', () => {
                 { distinctId: newUserDistinctId },
             ])
 
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personPropertyService({
                 event: '$pageview',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -830,7 +916,7 @@ describe('PersonState.update()', () => {
                 [{ distinctId: newUserDistinctId }]
             )
 
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personPropertyService({
                 event: '$pageview',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -872,19 +958,23 @@ describe('PersonState.update()', () => {
                 newUserUuid,
                 [{ distinctId: newUserDistinctId }]
             )
-
-            const personS = personState({
+            const event = {
                 event: '$pageview',
                 distinct_id: newUserDistinctId,
                 properties: {
                     $set_once: { c: 3, e: 4 },
                     $set: { b: 4 },
                 },
-            })
-            jest.spyOn(personS, 'handleIdentifyOrAlias').mockReturnValue(
+            }
+
+            // create mock merge service,
+            const mergeService = personMergeService(event)
+            jest.spyOn(mergeService, 'handleIdentifyOrAlias').mockReturnValue(
                 Promise.resolve([personInitial, Promise.resolve()])
             )
-            const [person, kafkaAcks] = await personS.update()
+
+            const personS = personProcessor(event, undefined, mergeService)
+            const [person, kafkaAcks] = await personS.processEvent()
             await hub.db.kafkaProducer.flush()
             await kafkaAcks
 
@@ -912,7 +1002,7 @@ describe('PersonState.update()', () => {
                 { distinctId: newUserDistinctId },
             ])
 
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personPropertyService({
                 event: '$pageview',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -948,12 +1038,18 @@ describe('PersonState.update()', () => {
                 { distinctId: newUserDistinctId },
             ])
 
-            const personS = personState({
-                event: '$pageview',
-                distinct_id: newUserDistinctId,
-                properties: {},
-            })
-            personS.updateIsIdentified = true
+            const personS = personPropertyService(
+                {
+                    event: '$pageview',
+                    distinct_id: newUserDistinctId,
+                    properties: {},
+                },
+                undefined,
+                true,
+                timestamp,
+                mainTeam,
+                true
+            )
 
             const [person, kafkaAcks] = await personS.updateProperties()
             await hub.db.kafkaProducer.flush()
@@ -977,8 +1073,6 @@ describe('PersonState.update()', () => {
             expect(persons.length).toEqual(1)
             expect(persons[0]).toEqual(person)
 
-            // Second call no-update
-            personS.updateIsIdentified = true // double just in case
             await personS.updateProperties()
             expect(hub.db.updatePersonDeprecated).toHaveBeenCalledTimes(1)
         })
@@ -1002,16 +1096,21 @@ describe('PersonState.update()', () => {
                 { distinctId: newUserDistinctId },
                 { distinctId: oldUserDistinctId },
             ]) // the merged Person
-            const personS = personState({
+
+            const event = {
                 event: '$pageview',
                 distinct_id: newUserDistinctId,
                 properties: { $set: { a: 7, d: 9 } },
-            })
-            jest.spyOn(personS, 'handleIdentifyOrAlias').mockReturnValue(
+            }
+            // create mock merge service
+            const mergeService = personMergeService(event)
+            jest.spyOn(mergeService, 'handleIdentifyOrAlias').mockReturnValue(
                 Promise.resolve([mergeDeletedPerson, Promise.resolve()])
             )
 
-            const [person, kafkaAcks] = await personS.update()
+            const personS = personProcessor(event, undefined, mergeService)
+
+            const [person, kafkaAcks] = await personS.processEvent()
             await hub.db.kafkaProducer.flush()
             await kafkaAcks
 
@@ -1038,7 +1137,7 @@ describe('PersonState.update()', () => {
 
     describe('on $identify event', () => {
         it(`no-op when $anon_distinct_id not passed`, async () => {
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personMergeService({
                 event: '$identify',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -1054,7 +1153,7 @@ describe('PersonState.update()', () => {
         })
 
         it(`creates person with both distinct_ids and marks user as is_identified when $anon_distinct_id passed`, async () => {
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personMergeService({
                 event: '$identify',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -1093,7 +1192,7 @@ describe('PersonState.update()', () => {
                 { distinctId: newUserDistinctId },
                 { distinctId: oldUserDistinctId },
             ])
-            const personS = personState({
+            const personS = personMergeService({
                 event: '$identify',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -1114,7 +1213,7 @@ describe('PersonState.update()', () => {
                     is_identified: false,
                 })
             )
-            expect(personS.updateIsIdentified).toBeTruthy()
+            expect(personS.getUpdateIsIdentified()).toBeTruthy()
 
             // verify Postgres persons
             const persons = await fetchPostgresPersonsH()
@@ -1126,7 +1225,7 @@ describe('PersonState.update()', () => {
             await createPerson(hub, timestamp, {}, {}, {}, teamId, null, false, newUserUuid, [
                 { distinctId: newUserDistinctId },
             ])
-            const personS = personState({
+            const personS = personMergeService({
                 event: '$identify',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -1148,7 +1247,7 @@ describe('PersonState.update()', () => {
                     is_identified: false,
                 })
             )
-            expect(personS.updateIsIdentified).toBeTruthy()
+            expect(personS.getUpdateIsIdentified()).toBeTruthy()
 
             // verify Postgres persons
             expect(persons.length).toEqual(1)
@@ -1164,7 +1263,7 @@ describe('PersonState.update()', () => {
                 { distinctId: oldUserDistinctId },
             ])
 
-            const personS = personState({
+            const personS = personMergeService({
                 event: '$identify',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -1187,7 +1286,7 @@ describe('PersonState.update()', () => {
                     is_identified: false,
                 })
             )
-            expect(personS.updateIsIdentified).toBeTruthy()
+            expect(personS.getUpdateIsIdentified()).toBeTruthy()
 
             // verify Postgres persons
             expect(persons.length).toEqual(1)
@@ -1206,7 +1305,7 @@ describe('PersonState.update()', () => {
                 { distinctId: newUserDistinctId },
             ])
 
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personMergeService({
                 event: '$identify',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -1273,7 +1372,7 @@ describe('PersonState.update()', () => {
                 { distinctId: newUserDistinctId },
             ])
 
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personMergeService({
                 event: '$identify',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -1340,7 +1439,7 @@ describe('PersonState.update()', () => {
                 { distinctId: newUserDistinctId },
             ])
 
-            const personS = personState({
+            const personS = personMergeService({
                 event: '$identify',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -1351,7 +1450,7 @@ describe('PersonState.update()', () => {
             await hub.db.kafkaProducer.flush()
             await kafkaAcks
 
-            expect(personS.updateIsIdentified).toBeTruthy()
+            expect(personS.getUpdateIsIdentified()).toBeTruthy()
             expect(person).toEqual(
                 expect.objectContaining({
                     id: expect.any(String),
@@ -1392,7 +1491,7 @@ describe('PersonState.update()', () => {
             await createPerson(hub, timestamp2, {}, {}, {}, teamId, null, true, newUserUuid, [
                 { distinctId: newUserDistinctId },
             ])
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personMergeService({
                 event: '$identify',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -1443,7 +1542,7 @@ describe('PersonState.update()', () => {
                 { distinctId: newUserDistinctId },
             ])
 
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personMergeService({
                 event: '$identify',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -1528,7 +1627,7 @@ describe('PersonState.update()', () => {
                 }
             )
 
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personMergeService({
                 event: '$identify',
                 distinct_id: oldUserDistinctId,
                 properties: {
@@ -1565,7 +1664,7 @@ describe('PersonState.update()', () => {
     describe('on $create_alias events', () => {
         // All the functionality tests are provided above in $identify tests, here we just make sure the calls are equivalent
         it('calls merge on $identify as expected', async () => {
-            const state: PersonState = personState(
+            const state: PersonMergeService = personMergeService(
                 {
                     event: '$identify',
                     distinct_id: newUserDistinctId,
@@ -1582,7 +1681,7 @@ describe('PersonState.update()', () => {
         })
 
         it('calls merge on $create_alias as expected', async () => {
-            const state: PersonState = personState(
+            const state: PersonMergeService = personMergeService(
                 {
                     event: '$create_alias',
                     distinct_id: newUserDistinctId,
@@ -1600,7 +1699,7 @@ describe('PersonState.update()', () => {
         })
 
         it('calls merge on $merge_dangerously as expected', async () => {
-            const state: PersonState = personState(
+            const state: PersonMergeService = personMergeService(
                 {
                     event: '$merge_dangerously',
                     distinct_id: newUserDistinctId,
@@ -1627,7 +1726,7 @@ describe('PersonState.update()', () => {
             await createPerson(hub, timestamp2, {}, {}, {}, teamId, null, true, newUserUuid, [
                 { distinctId: newUserDistinctId },
             ])
-            const [person, kafkaAcks] = await personState({
+            const [person, kafkaAcks] = await personMergeService({
                 event: '$merge_dangerously',
                 distinct_id: newUserDistinctId,
                 properties: {
@@ -1690,7 +1789,7 @@ describe('PersonState.update()', () => {
     describe('illegal aliasing', () => {
         const illegalIds = ['', '   ', 'null', 'undefined', '"undefined"', '[object Object]', '"[object Object]"']
         it.each(illegalIds)('stops $identify if current distinct_id is illegal: `%s`', async (illegalId: string) => {
-            const [person] = await personState({
+            const [person] = await personMergeService({
                 event: '$identify',
                 distinct_id: illegalId,
                 properties: {
@@ -1704,7 +1803,7 @@ describe('PersonState.update()', () => {
         })
 
         it.each(illegalIds)('stops $identify if $anon_distinct_id is illegal: `%s`', async (illegalId: string) => {
-            const [person] = await personState({
+            const [person] = await personMergeService({
                 event: '$identify',
                 distinct_id: 'some_distinct_id',
                 properties: {
@@ -1718,7 +1817,7 @@ describe('PersonState.update()', () => {
         })
 
         it('stops $create_alias if current distinct_id is illegal', async () => {
-            const [person] = await personState({
+            const [person] = await personMergeService({
                 event: '$create_alias',
                 distinct_id: 'false',
                 properties: {
@@ -1732,7 +1831,7 @@ describe('PersonState.update()', () => {
         })
 
         it('stops $create_alias if alias is illegal', async () => {
-            const [person] = await personState({
+            const [person] = await personMergeService({
                 event: '$create_alias',
                 distinct_id: 'some_distinct_id',
                 properties: {
@@ -1790,14 +1889,14 @@ describe('PersonState.update()', () => {
 
             // this event means the person will be merged
             // so hashkeyoverride should be updated to the new person id whichever way we merged
-            await personState({
+            await personProcessor({
                 event: '$identify',
                 distinct_id: 'new_distinct_id',
                 properties: {
                     $anon_distinct_id: 'anonymous_id',
                     distinct_id: 'new_distinct_id',
                 },
-            }).update()
+            }).processEvent()
             await hub.db.kafkaProducer.flush()
 
             const [person] = await fetchPostgresPersonsH()
@@ -1878,14 +1977,14 @@ describe('PersonState.update()', () => {
             // this event means the person will be merged
             // so hashkeyoverride should be updated to be either
             // we're optimizing on updates to not write on conflict and ordering is not guaranteed
-            await personState({
+            await personProcessor({
                 event: '$identify',
                 distinct_id: 'new_distinct_id',
                 properties: {
                     $anon_distinct_id: 'anonymous_id',
                     distinct_id: 'new_distinct_id',
                 },
-            }).update()
+            }).processEvent()
             await hub.db.kafkaProducer.flush()
 
             const [person] = await fetchPostgresPersonsH()
@@ -1954,13 +2053,13 @@ describe('PersonState.update()', () => {
                 hash_key: 'different_id',
             })
 
-            await personState({
+            await personProcessor({
                 event: '$identify',
                 distinct_id: 'new_distinct_id',
                 properties: {
                     $anon_distinct_id: 'anonymous_id',
                 },
-            }).update()
+            }).processEvent()
             await hub.db.kafkaProducer.flush()
 
             const [person] = await fetchPostgresPersonsH()
@@ -2012,7 +2111,7 @@ describe('PersonState.update()', () => {
                 { distinctId: secondUserDistinctId },
             ])
 
-            const state: PersonState = personState({}, hub)
+            const state: PersonMergeService = personMergeService({}, hub)
             jest.spyOn(hub.db.kafkaProducer, 'queueMessages')
             const [person, kafkaAcks] = await state.merge(secondUserDistinctId, firstUserDistinctId, teamId, timestamp)
             await hub.db.kafkaProducer.flush()
@@ -2041,7 +2140,7 @@ describe('PersonState.update()', () => {
                 { distinctId: secondUserDistinctId },
             ])
 
-            const state: PersonState = personState({}, hub)
+            const state: PersonMergeService = personMergeService({}, hub)
             jest.spyOn(hub.db.kafkaProducer, 'queueMessages')
             const [person, kafkaAcks] = await state.mergePeople({
                 mergeInto: first,
@@ -2108,7 +2207,7 @@ describe('PersonState.update()', () => {
             const second = await createPerson(hub, timestamp, {}, {}, {}, teamId, null, false, secondUserUuid, [
                 { distinctId: secondUserDistinctId },
             ])
-            const state: PersonState = personState({}, hub)
+            const state: PersonMergeService = personMergeService({}, hub)
             // break postgres
             const error = new DependencyUnavailableError('testing', 'Postgres', new Error('test'))
             jest.spyOn(hub.db.postgres, 'transaction').mockImplementation(() => {
@@ -2159,7 +2258,7 @@ describe('PersonState.update()', () => {
             await createPerson(hub, timestamp, {}, {}, {}, teamId, null, false, secondUserUuid, [
                 { distinctId: secondUserDistinctId },
             ])
-            const state: PersonState = personState({}, hub)
+            const state: PersonMergeService = personMergeService({}, hub)
             // break postgres
             const error = new DependencyUnavailableError('testing', 'Postgres', new Error('test'))
             jest.spyOn(state, 'mergePeople').mockImplementation(() => {
@@ -2208,7 +2307,7 @@ describe('PersonState.update()', () => {
                 { distinctId: secondUserDistinctId },
             ])
 
-            const state: PersonState = personState(
+            const state: PersonMergeService = personMergeService(
                 {
                     event: '$merge_dangerously',
                     distinct_id: firstUserDistinctId,
@@ -2295,7 +2394,7 @@ describe('JSONB optimization flag compatibility', () => {
             { distinctId: legacyDistinctId },
         ])
 
-        const legacyPersonState = new PersonState(
+        const personContext = new PersonContext(
             { team_id: teamId, properties: {}, ...eventData, distinct_id: legacyDistinctId } as any,
             mainTeam,
             legacyDistinctId,
@@ -2306,8 +2405,9 @@ describe('JSONB optimization flag compatibility', () => {
             0,
             0.0
         )
+        const legacyPropertyService = new PersonPropertyService(personContext)
 
-        const [legacyPerson, legacyKafkaAcks] = await legacyPersonState.updateProperties()
+        const [legacyPerson, legacyKafkaAcks] = await legacyPropertyService.updateProperties()
         await hub.db.kafkaProducer.flush()
         await legacyKafkaAcks
 
@@ -2315,7 +2415,7 @@ describe('JSONB optimization flag compatibility', () => {
             { distinctId: optimizedDistinctId },
         ])
 
-        const optimizedPersonState = new PersonState(
+        const optimizedPersonContext = new PersonContext(
             { team_id: teamId, properties: {}, ...eventData, distinct_id: optimizedDistinctId } as any,
             mainTeam,
             optimizedDistinctId,
@@ -2326,8 +2426,9 @@ describe('JSONB optimization flag compatibility', () => {
             0,
             1.0
         )
+        const optimizedPropertyService = new PersonPropertyService(optimizedPersonContext)
 
-        const [optimizedPerson, optimizedKafkaAcks] = await optimizedPersonState.updateProperties()
+        const [optimizedPerson, optimizedKafkaAcks] = await optimizedPropertyService.updateProperties()
         await hub.db.kafkaProducer.flush()
         await optimizedKafkaAcks
 
@@ -2498,7 +2599,7 @@ describe('JSONB optimization flag compatibility', () => {
                 [{ distinctId: `${oldUserDistinctId}-${uuid_suffix}` }]
             )
 
-            const legacyPersonState = new PersonState(
+            const legacyPersonContext = new PersonContext(
                 { team_id: teamId, properties: {}, event: '$pageview', distinct_id: newUserDistinctId } as any,
                 mainTeam,
                 newUserDistinctId,
@@ -2509,9 +2610,10 @@ describe('JSONB optimization flag compatibility', () => {
                 0,
                 0.0
             )
-            legacyPersonState.updateIsIdentified = updateIsIdentified
+            legacyPersonContext.updateIsIdentified = updateIsIdentified
+            const legacyPropertyService = new PersonPropertyService(legacyPersonContext)
 
-            const optimizedPersonState = new PersonState(
+            const optimizedPersonContext = new PersonContext(
                 { team_id: teamId, properties: {}, event: '$pageview', distinct_id: oldUserDistinctId } as any,
                 mainTeam,
                 oldUserDistinctId,
@@ -2522,10 +2624,11 @@ describe('JSONB optimization flag compatibility', () => {
                 0,
                 1.0
             )
-            optimizedPersonState.updateIsIdentified = updateIsIdentified
+            optimizedPersonContext.updateIsIdentified = updateIsIdentified
+            const optimizedPropertyService = new PersonPropertyService(optimizedPersonContext)
 
-            const [legacyResult] = await legacyPersonState.updateProperties()
-            const [optimizedResult] = await optimizedPersonState.updateProperties()
+            const [legacyResult] = await legacyPropertyService.updateProperties()
+            const [optimizedResult] = await optimizedPropertyService.updateProperties()
 
             expect(legacyResult.is_identified).toEqual(optimizedResult.is_identified)
             expect(legacyResult.properties).toEqual(optimizedResult.properties)
